# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: MyAgent API
  description: |
    AI Agent 管理平台后端 API。

    ## 认证方式

    除健康检查和认证端点外，所有 API 请求需在 `Authorization` 头携带 Bearer token：

    ```
    Authorization: Bearer <access_token>
    ```

    Token 通过 `/api/auth/login` 或 `/api/auth/register` 获取。

    ## 通用响应格式

    所有标准 API 响应遵循统一信封格式：
    - 成功：`{ "success": true, "data": T, "meta?": { total, page, limit } }`
    - 失败：`{ "success": false, "error": string, "message": string }`

    错误 HTTP 状态码对照表：

    | 状态码 | 含义 | 前端行为 |
    |--------|------|---------|
    | 401 | 未认证 / Token 过期 | 拦截器自动尝试 refresh token |
    | 403 | 无权限 | 跳转 403 页面 |
    | 404 | 资源不存在 | 显示 404 |
    | 409 | 资源冲突 | 提示用户 |
    | 422 | 参数校验失败 | 显示字段级错误 |
    | 500 | 服务器内部错误 | 显示通用错误页 |
  version: "0.2.0"
  contact:
    name: MyAgent Team
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
  - url: http://localhost:3000
    description: 前端 Next.js 开发服务器（通过代理转发）
paths:
...
  /api/auth/register:
    post:
      tags: [认证]
      summary: 用户注册
      description: 注册新用户，成功后返回 JWT token 对和用户信息（自动登录）。
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: 注册成功，返回 token 和用户信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: 邮箱已被注册
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error409"
        "422":
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error422"
  /api/auth/login:
    post:
      tags: [认证]
      summary: 用户登录
      description: 使用邮箱和密码登录，返回 JWT token 对和用户信息。
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: 邮箱或密码错误 / 账户被禁用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error401"
...
  /api/auth/refresh:
    post:
      tags: [认证]
      summary: 刷新 Token
      description: 使用 refresh_token 获取新的 access_token 和 refresh_token。
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRefreshRequest"
      responses:
        "200":
          description: Token 刷新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenRefreshResponse"
        "401":
          description: refresh_token 无效或已过期
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error401"
  /api/auth/profile:
    get:
      tags: [认证]
      summary: 获取个人信息
      description: 获取当前登录用户的详细信息。
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 用户信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/auth/password:
    put:
      tags: [认证]
      summary: 修改密码
      description: 修改当前登录用户的密码。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeRequest"
      responses:
        "200":
          description: 密码修改成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: 原密码错误或新密码不符合要求
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error400"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

  # ============================================================================
  # Health
  # ============================================================================
...
  /agents:
    get:
      tags: [Agent 管理]
      summary: Agent 列表
      description: 获取 Agent 列表，支持分页和搜索。
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: 页码
        - name: search
          in: query
          schema:
            type: string
          description: 搜索关键词（匹配名称和描述）
      responses:
        "200":
          description: Agent 列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessWithData"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Agent"
                      total:
                        type: integer
                        example: 42
        "401":
          $ref: "#/components/responses/UnauthorizedError"

    post:
      tags: [Agent 管理]
      summary: 创建 Agent
      description: 创建一个新的 Agent。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentCreateRequest"
      responses:
        "201":
          description: Agent 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/Agent"
        "422":
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error422"
...
  /knowledge/files:
    get:
      tags: [知识库]
      summary: 知识文档列表
      description: 获取知识库中所有文档的列表。
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 文档列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/KnowledgeDocument"

    post:
      tags: [知识库]
      summary: 上传知识文档
      description: 上传文件到知识库（multipart/form-data）。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: 要上传的文件
              required:
                - file
      responses:
        "201":
          description: 上传成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/KnowledgeDocument"
        "413":
          description: 文件过大
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error413"
  /knowledge/files/{id}:
    delete:
      tags: [知识库]
      summary: 删除知识文档
      description: 删除指定知识文档及其向量索引。
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 删除成功（无内容）
        "404":
          description: 文档不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error404"
  /knowledge/search:
    post:
      tags: [知识库]
      summary: 语义搜索
      description: 对知识库进行语义搜索。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: 搜索查询
                limit:
                  type: integer
                  default: 10
                  minimum: 1
                  maximum: 100
              required:
                - query
      responses:
        "200":
          description: 搜索结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SearchResult"

  # ============================================================================
  # Memory
  # ============================================================================
  /memory:
    get:
      tags: [记忆]
      summary: 记忆列表
      description: 获取当前用户的记忆列表，支持分页。
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: 页码
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/MemoryType"
          description: 按记忆类型筛选
      responses:
        "200":
          description: 记忆列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Memory"
  /memory/search:
    post:
      tags: [记忆]
      summary: 语义搜索记忆
      description: 对记忆进行语义搜索。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemorySearchRequest"
      responses:
        "200":
          description: 搜索结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Memory"
...
  /tools:
    get:
      tags: [工具]
      summary: 工具列表
      description: 获取所有可用工具的列表。
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 工具列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tool"
...
  /logs:
    get:
      tags: [日志]
      summary: 日志列表
      description: 获取系统日志，支持按级别、搜索、来源等条件筛选。
      security:
        - BearerAuth: []
      parameters:
        - name: level
          in: query
          schema:
            type: string
            enum: [DEBUG, INFO, WARN, ERROR]
          description: 日志级别
        - name: search
          in: query
          schema:
            type: string
          description: 搜索关键词（匹配消息内容）
        - name: source
          in: query
          schema:
            type: string
          description: 日志来源
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: 页码
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
          description: 每页数量
      responses:
        "200":
          description: 日志列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/LogEntry"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"

  # ============================================================================
  # Admin
  # ============================================================================
  /admin/users:
    get:
      tags: [管理后台]
      summary: 用户列表（管理）
      description: 获取所有用户列表（仅管理员）。
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 用户列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminUser"
        "403":
          $ref: "#/components/responses/ForbiddenError"
...
  /admin/system:
    get:
      tags: [管理后台]
      summary: 系统指标
      description: 获取系统运行指标（仅管理员）。
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 系统指标
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/SystemMetrics"
        "403":
          $ref: "#/components/responses/ForbiddenError"

    put:
      tags: [管理后台]
      summary: 更新系统配置
      description: 更新系统级配置（仅管理员）。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: 系统配置键值对
              example:
                max_upload_size_mb: 50
                allow_registration: true
      responses:
        "200":
          description: 配置更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          $ref: "#/components/responses/ForbiddenError"

  # ============================================================================
  # Settings
  # ============================================================================
  /settings:
    get:
      tags: [设置]
      summary: 获取设置
      description: 获取当前用户的个人设置。
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 用户设置
          content:
            application/json:
              schema:
                type: object
                properties:
                  theme:
                    type: string
                    enum: [light, dark, system]
                    example: system
                  language:
                    type: string
                    example: zh-CN
                  notifications:
                    type: boolean
                    example: true

    put:
      tags: [设置]
      summary: 更新设置
      description: 更新当前用户的个人设置。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                theme:
                  type: string
                  enum: [light, dark, system]
                language:
                  type: string
                notifications:
                  type: boolean
      responses:
        "200":
          description: 设置更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ============================================================================
  # Dashboard
  # ============================================================================
  /dashboard/overview:
    get:
      tags: [仪表盘]
      summary: 总览数据
      description: 获取仪表盘总览数据。
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 总览数据
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_messages:
                    type: integer
                    example: 1523
                  total_sessions:
                    type: integer
                    example: 89
                  total_agents:
                    type: integer
                    example: 12
                  avg_response_time:
                    type: number
                    format: float
                    example: 1.25
                    description: 平均响应时间（秒）
  /dashboard/recent-chats:
    get:
      tags: [仪表盘]
      summary: 最近聊天
      description: 获取最近的聊天记录。
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 最近聊天列表
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    title:
                      type: string
                    message_count:
                      type: integer
                    created_at:
                      type: string
                      format: date-time
  /dashboard/recent-agents:
    get:
      tags: [仪表盘]
      summary: 最近 Agent
      description: 获取最近使用的 Agent 列表。
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 最近 Agent 列表
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    description:
                      type: string
                    status:
                      $ref: "#/components/schemas/AgentStatus"
                    updated_at:
                      type: string
                      format: date-time
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token，通过登录或注册获取
  schemas:
...
    Error401:
      type: object
      properties:
        detail:
          type: string
          example: 邮箱或密码错误
...
    Error409:
      type: object
      properties:
        detail:
          type: string
          example: 该邮箱已被注册
...
    Error422:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
                example: ["body", "email"]
              msg:
                type: string
                example: Value error, 邮箱格式不正确
              type:
                type: string
                example: value_error
...
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 50
          description: 用户名
          example: 张三
        email:
          type: string
          format: email
          description: 邮箱
          example: zhangsan@example.com
        password:
          type: string
          minLength: 6
          description: 密码
          example: password123
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: zhangsan@example.com
        password:
          type: string
          example: password123
    TokenRefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: 刷新令牌
    TokenRefreshResponse:
      type: object
      properties:
        access_token:
          type: string
          description: 新的 access_token
        refresh_token:
          type: string
          description: 新的 refresh_token
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: 张三
        email:
          type: string
          example: zhangsan@example.com
        avatar:
          type: string
          default: ""
          example: https://example.com/avatar.png
        role:
          type: string
          enum: [USER, ADMIN]
          example: USER
        is_active:
          type: boolean
          default: true
        created_at:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access_token（有效期 24 小时）
          example: eyJhbGciOiJIUzI1NiIs...
        refresh_token:
          type: string
          description: JWT refresh_token（有效期 7 天）
          example: eyJhbGciOiJIUzI1NiIs...
        user:
          $ref: "#/components/schemas/UserResponse"

    # ── Chat ────────────────────────────────────────────────────
...
  responses:
    UnauthorizedError:
      description: 未认证或 Token 过期
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error401"
...